backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
anthropic==0.18.0
openai==1.12.0  # for Whisper if needed

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Integrations
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
plaid-python==16.0.0
twilio==8.13.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.6

# CORS & Security
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
